requirements.txt
fastapi
uvicorn[standard]
numpy<2
Pillow
python-multipart
jinja2
onnxruntime
